Moved square logic to a custom hook. Refactored Home component to use the hook for managing squares and buttons.
