+++ src/scss/main.scss
@@ -1,3 +1,4 @@
 @import "./base/base";
 @import "./views/home";
 @import "./components/squares";
+@import "./components/buttons";
