+++ src/views/Home.js
@@ -1,15 +1,37 @@
 import Squares from "../components/Squares";
 import Button from "../components/Button";
+import { UseSquares } from "../hooks/UseSquares";
 
 const Home = () => {
+  const {
+    squares,
+    gridSize,
+    addSquare,
+    clearSquares,
+    squareStyles,
+    gridStyle,
+  } = UseSquares();
   return (
-    <section>
+    <section className="squares-interaction">
       <div className="button-container">
-        <Button onClick={() => {}} label="Add Square" />
-        <Button onClick={() => {}} label="Clear All Squares" />
+        <Button
+          onClick={addSquare}
+          label="Add Square"
+          className="add-square-button"
+        />
+        <Button
+          onClick={clearSquares}
+          label="Clear All Squares"
+          className="clear-squares-button"
+        />
       </div>
       <div className="container">
-        <Squares />
+        <Squares
+          squares={squares}
+          gridSize={gridSize}
+          squareStyles={squareStyles}
+          gridStyle={gridStyle}
+        />
       </div>
     </section>
   );
